ensure we load float32 into redis

+++ cache_dict_to_redis.py
@@ -22,7 +22,7 @@ def save_to_redis(vector_cache, cache):
     idx = 0
     print(f"Saving {len(cache)}")
     for passage, vector in cache.items():
-        vector_cache.set(passage, np.array(vector))
+        vector_cache.set(passage, np.array(vector, dtype=np.float32))
         if (idx % 10000 == 0):
             print(f"Saved {idx}")
         idx += 1
@@ -31,4 +31,4 @@ def save_to_redis(vector_cache, cache):
 if __name__ == "__main__":
     r = redis.Redis(host='localhost', port=6379)
     cache = load_encode_cache()
-    save_to_redis(VectorCache(r), cache)
+    save_to_redis(VectorCache(r, dtype=np.float32), cache)
